+++ mlrun/api/alembic.ini
@@ -35,7 +35,7 @@ script_location = alembic
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# this is overriden anyway in the alembic env.py by what ever is configured in mlrun.mlconf
+# this is overridden anyway in the alembic env.py by what ever is configured in mlrun.mlconf
 sqlalchemy.url = sqlite:////mlrun/db/mlrun.db
