--- mlrun/api/alembic/versions/79021d71037e_schedule_labels.py
@@ -1,41 +1,0 @@
-"""schedule labels
-
-Revision ID: 79021d71037e
-Revises: 11f8dd2dc9fe
-Create Date: 2020-10-06 16:42:57.581736
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "79021d71037e"
-down_revision = "11f8dd2dc9fe"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "schedules_v2_labels",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("value", sa.String(), nullable=True),
-        sa.Column("schedule_project", sa.Integer(), nullable=True),
-        sa.Column("schedule_name", sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(["schedule_name"], ["schedules_v2.name"],),
-        sa.ForeignKeyConstraint(["schedule_project"], ["schedules_v2.project"],),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint(
-            "name", "schedule_project", "schedule_name", name="_schedules_v2_labels_uc"
-        ),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("schedules_v2_labels")
-    # ### end Alembic commands ###
