+++ mlrun/api/alembic/versions/bf1a9b79abf9_schedule_labels.py
@@ -1,0 +1,41 @@
+"""schedule labels
+
+Revision ID: bf1a9b79abf9
+Revises: 11f8dd2dc9fe
+Create Date: 2020-10-07 10:09:06.324628
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'bf1a9b79abf9'
+down_revision = '11f8dd2dc9fe'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        'schedules_v2_labels',
+        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('name', sa.String(), nullable=True),
+        sa.Column('value', sa.String(), nullable=True),
+        sa.Column('project', sa.Integer(), nullable=True),
+        sa.Column('schedule_name', sa.Integer(), nullable=True),
+        sa.ForeignKeyConstraint(['project'], ['schedules_v2.project'],),
+        sa.ForeignKeyConstraint(['schedule_name'], ['schedules_v2.name'],),
+        sa.PrimaryKeyConstraint('id'),
+        sa.UniqueConstraint(
+            'name', 'project', 'schedule_name', name='_schedules_v2_labels_uc'
+        ),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('schedules_v2_labels')
+    # ### end Alembic commands ###
